+++ src/main/java/roomescape/controller/admin/AdminTimeController.java
@@ -1,0 +1,39 @@
+package roomescape.controller.admin;
+
+import jakarta.validation.Valid;
+import java.net.URI;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.service.ReservationTimeService;
+import roomescape.service.dto.request.ReservationTimeRequest;
+import roomescape.service.dto.response.ReservationTimeResponse;
+
+@RestController
+@RequestMapping("/admin/times")
+public class AdminTimeController {
+
+    private final ReservationTimeService reservationTimeService;
+
+    public AdminTimeController(ReservationTimeService reservationTimeService) {
+        this.reservationTimeService = reservationTimeService;
+    }
+
+    @PostMapping
+    public ResponseEntity<ReservationTimeResponse> addTime(
+            @RequestBody @Valid ReservationTimeRequest reservationTimeRequest) {
+        ReservationTimeResponse reservationTimeResponse = reservationTimeService.save(reservationTimeRequest);
+        return ResponseEntity.created(URI.create("/admin/time"))
+                .body(reservationTimeResponse);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteTime(@PathVariable Long id) {
+        reservationTimeService.deleteById(id);
+        return ResponseEntity.noContent().build();
+    }
+}
